requirements.txt
flask>=2.3.0
cryptography>=41.0.0
gunicorn>=21.0.0
requests>=2.31.0
kengerkit>=0.2.0
pymysql>=1.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
schedule>=1.2.0
